move idl for management canister to assets
